refactor(data): move organised-workshops JSON-LD into a data module
~100 lines of static schema.org Event graph were hardcoded in the
conference-activity route component. Move it to
$lib/data/organisedWorkshops.ts (importing author/website from
siteConfig there). The object literal is unchanged, so JSON.stringify
output is byte-identical; the route now just imports the string and
feeds it to useJsonLdScript.
